스프링 시큐리티 적용, 예외처리

+++ src/main/java/com/toy/firstduoproject/controller/PostController.java
@@ -1,14 +1,19 @@
 package com.toy.firstduoproject.controller;
 
+import com.toy.firstduoproject.config.auth.PrincipalDetails;
 import com.toy.firstduoproject.domain.entity.Posts;
+import com.toy.firstduoproject.handler.ex.MemberDifferentException;
+import com.toy.firstduoproject.handler.ex.PostEmptyException;
 import com.toy.firstduoproject.service.PostService;
 import com.toy.firstduoproject.service.dto.PostSaveRequestDto;
 import com.toy.firstduoproject.service.dto.PostUpdateRequestDto;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @Controller
@@ -29,9 +34,18 @@ public class PostController {
     }
 
     @PostMapping("/posts/add")
-    public String addPost(@ModelAttribute PostSaveRequestDto requestDto) {
-        postService.createPost(requestDto);
-        return "redirect:/posts";
+    public String addPost(@Valid PostSaveRequestDto requestDto, @AuthenticationPrincipal PrincipalDetails principalDetails) {
+        Long id = 0L;
+        try {
+            PostSaveRequestDto saveRequestDto = new PostSaveRequestDto(requestDto.getTitle(),
+                    requestDto.getContent(),
+                    principalDetails.getMember());
+            Posts post = postService.createPost(saveRequestDto);
+            id = post.getId();
+        }catch (Exception e){
+            throw new PostEmptyException("제목 혹은 내용이 비어있습니다.");
+        }
+        return "redirect:/posts/"+id;
     }
 
     @GetMapping("/posts/{post-id}")
@@ -49,10 +63,17 @@ public class PostController {
     }
 
     @GetMapping("/posts/{post-id}/edit")
-    public String editForm(@PathVariable("post-id") Long postId, Model model){
+    public String editForm(@PathVariable("post-id") Long postId, Model model,@AuthenticationPrincipal PrincipalDetails principalDetails){
+
         Posts post = postService.findById(postId);
+
+        if(post.getMember().getId()!=principalDetails.getMember().getId()){
+            throw new MemberDifferentException("글을 수정할 권리가 없습니다.");
+        };
+
         model.addAttribute("post",post);
         return "posts/editPost";
+
     }
 
     @PostMapping("/posts/{post-id}/edit")
@@ -63,8 +84,15 @@ public class PostController {
     }
 
     @PostMapping("/posts/{post-id}")
-    public String deletePost(@PathVariable("post-id") Long postId){
-        postService.deletePost(postId);
-        return "redirect:/posts";
+    public String deletePost(@PathVariable("post-id") Long postId,@AuthenticationPrincipal PrincipalDetails principalDetails){
+
+        Posts post = postService.findById(postId);
+
+        if(post.getMember().getId()==principalDetails.getMember().getId()||principalDetails.getMember().getRole().equals("ROLE_ADMIN")){
+            postService.deletePost(postId);
+            return "redirect:/posts";
+        };
+
+        throw new MemberDifferentException("글을 수정할 권리가 없습니다.");
     }
 }
